### 2023-07-14 #5660 @editorialbot
[PRE REVIEW]: Phylotrack: C++ and Python libraries for in silico phylogenetic tracking
**Submitting author:** @emilydolson (<a href="http://orcid.org/0000-0001-8616-4898">Emily Dolson</a>)
**Repository:** https://github.com/emilydolson/phylotrackpy
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.13
