+++ systematics_bindings.cpp
@@ -233,14 +233,14 @@ PYBIND11_MODULE(systematics, m) {
         )mydelimiter")
         .def("get_num_outside", static_cast<size_t (sys_t::*) () const>(&sys_t::GetNumOutside), R"mydelimiter(
             Returns the number of *extinct* taxa that have no living descendants.
-            By definition, these are the taxa that are not in the current phylogeny.
+            By definition, these are the taxa that are not in the current phylogenetic tree.
         )mydelimiter")
         .def("get_tree_size", static_cast<size_t (sys_t::*) () const>(&sys_t::GetTreeSize), R"mydelimiter(
-            Returns the number of taxa in the current phylogeny.
+            Returns the number of taxa in the current phylogenetic tree.
         )mydelimiter")
         .def("get_num_taxa", static_cast<size_t (sys_t::*) () const>(&sys_t::GetNumTaxa), R"mydelimiter(
             Returns the number of total taxa.
-            Empirically, this is equal to the number of taxa in the current phylogeny, plus the number of extinct taxa with no living descendants.
+            Empirically, this is equal to the number of taxa in the current phylogenetic tree, plus the number of extinct taxa with no living descendants.
         )mydelimiter")
         .def("get_max_depth", static_cast<int (sys_t::*) () >(&sys_t::GetMaxDepth), R"mydelimiter(
             Returns the lineage length (phylogenetic depth) of the active taxon with the longest lineage.
@@ -444,18 +444,18 @@ PYBIND11_MODULE(systematics, m) {
             If no MRCA exists, this returns -1.
         )mydelimiter")
         .def("colless_like_index", static_cast<double (sys_t::*) () const>(&sys_t::CollessLikeIndex), R"mydelimiter(
-            Calculates and returns the Colless-like Index of the currently-active phylogeny. This metric is used to measure tree balance in multifurcating
+            Calculates and returns the Colless-like Index of the currently-active phylogenetic tree. This metric is used to measure tree balance in multifurcating
             trees, as described in doi:10.1371/journal.pone.0203401.
             Most people are familiar with the standard Colless Index. However, that version is only applicable to bifurcating trees.
         )mydelimiter")
         .def("sackin_index", static_cast<int (sys_t::*) () const>(&sys_t::SackinIndex), R"mydelimiter(
-            Calculates and returns the Sackin Index of the currently-active phylogeny. This metric is used to measure phylogenetic tree balance.
+            Calculates and returns the Sackin Index of the currently-active phylogenetic tree. This metric is used to measure phylogenetic tree balance.
             It is calcualted by adding together the depth of every leaf to its Most-Recent Common Ancestor (or its subtree root if none is found).
             For more information, see doi:10.2307/2992186.
         )mydelimiter")
         .def("get_branches_to_root", [](sys_t & self, taxon_t * tax){return self.GetBranchesToRoot(tax);}, R"mydelimiter(
             Given a taxon, this function calculates and returns the number of branches leading to multiple extant taxa on the path to its Most-Recent
-            Common Ancestor (or its subtree root if none is found). Only extant taxa are considered since most phylogeny reconstruction algorithms are not designed to additionaly handle extinct taxa. Moreover, this function also does not count unifurcations -- that is, points at which each taxon
+            Common Ancestor (or its subtree root if none is found). Only extant taxa are considered since most phylogenetic tree reconstruction algorithms are not designed to additionaly handle extinct taxa. Moreover, this function also does not count unifurcations -- that is, points at which each taxon
             only has a single ancestor.
 
             Parameters
@@ -464,29 +464,50 @@ PYBIND11_MODULE(systematics, m) {
[…]
