+++ joss/paper.bib
@@ -176,6 +176,7 @@
 	journaltitle = {Artificial Life},
 	author = {Dolson, Emily and Lalejini, Alexander and Jorgensen, Steven and Ofria, Charles},
 	date = {2020},
+  doi = {10.1162/artl_a_00313}
 }
 
 
@@ -314,7 +315,8 @@ year = {2019}
   number={14},
   pages={2079--2080},
   year={2014},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi = {10.1093/bioinformatics/btu157}
 }
 
 @article{sukumaran2010dendropy,
@@ -325,7 +327,8 @@ year = {2019}
   number={12},
   pages={1569--1571},
   year={2010},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi = {10.1093/bioinformatics/btq228}
 }
 
 @article{cock2009biopython,
@@ -336,7 +339,8 @@ year = {2019}
   number={11},
   pages={1422--1423},
   year={2009},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi = {10.1093/bioinformatics/btp163}
 }
 
 @article{challa2019phylogenetic,
@@ -345,7 +349,8 @@ year = {2019}
   journal={Essentials of Bioinformatics, Volume III: In Silico Life Sciences: Agriculture},
   pages={167--192},
   year={2019},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.1007/978-3-030-19318-8_10}
 }
 
 @article{emerson2008phylogenetic,
@@ -356,7 +361,8 @@ year = {2019}
   number={11},
   pages={619--630},
   year={2008},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/j.tree.2008.07.005}
 }
 
 @article{stadler2011simulating,
