### 2023-07-17 132825b
Add "new" to get_num_offspring
