### 2023-07-17 89ada94
Fix get_offspring
