### 2023-07-17 e002ba5
Add warning to set_store_active
