### 2023-07-17 88a1258
Add warning to set_store_ancestors
