### 2023-07-17 79290f1
Add more warnings to set_store_active
