### 2023-07-17 619cf28
Add warning to set_store_archive
