### 2023-07-17 72a7f25
Fix get_store_position
