### 2023-07-17 827ca45
Add warn to set_store_outside
