### 2023-07-17 db72793
Added citation get_phylo_diversity
