### 2023-07-17 8cf07f8
Drop "moreover" from get_branches_to_root
