### 2023-07-17 ed67cb2
Added Shannon Diversity note to calc_diversity
