### 2023-07-17 0d188fb
Fix get_next_parent
