### 2023-07-17 116d7af
Simplify reference docstring.
