### 2023-07-17 d2b8051
Fix get_max_depth
