+++ systematics_bindings.cpp
@@ -55,28 +55,17 @@ PYBIND11_MODULE(systematics, m) {
         .def("get_pop_ID", &emp::WorldPosition::GetPopID, R"mydelimiter(
             Returns the ID (as an int) of the population that this WorldPosition is referring to.
 
-            Wondering why there might be multiple populations? It's because WorldPosition objects
-            are designed to gracefully handle configurations in which there are
-            multiple separate populations (e.g. due to islands or separated generations).
+            Wondering why there might be multiple populations? It's because WorldPosition objects are designed to gracefully handle configurations in which there are multiple separate populations (e.g. due to islands or separated generations).
             If you're just using one population, the pop ID will always be 0 and that's fine.
             )mydelimiter")
         .def("is_active", &emp::WorldPosition::IsActive, R"mydelimiter(
-            Returns a boolean indicating whether this position potentially represents an "active" (i.e. "alive")
-            organism in the context of a generational/synchronous configuration (i.e. one in which a new generation
-            is created on each time step and there is no overlap of organisms between generations; this
-            configuration is commonly used in evolutionary computation and simple evolutionary models). In this case,
-            the next generation (i.e. the one that is currently in the process of being created) is considered
-            "inactive". Conventionally, the population that selection is currently happening on has ID 0 and the population
-            holding the new generation has ID 1. Thus, this method simply returns true if the population ID is 0 and false
-            otherwise.
+            Returns a boolean indicating whether this position potentially represents an "active" (i.e. "alive") organism in the context of a generational/synchronous configuration (i.e. one in which a new generation is created on each time step and there is no overlap of organisms between generations; this configuration is commonly used in evolutionary computation and simple evolutionary models). In this case, the next generation (i.e. the one that is currently in the process of being created) is considered "inactive". Conventionally, the population that selection is currently happening on has ID 0 and the population holding the new generation has ID 1. Thus, this method simply returns true if the population ID is 0 and false otherwise.
 
-            If you are using population IDs for a separate purpose (e.g. islands), this method will not yield accurate
-            results.
+            If you are using population IDs for a separate purpose (e.g. islands), this method will not yield accurate results.
             )mydelimiter")
         .def("is_valid", &emp::WorldPosition::IsValid, R"mydelimiter(
             -1 can be used as a sentinel value to indicate that a position is not to be used.
-            This method returns a boolean indicating whether this position is "valid"
-            (i.e. the index is not equal to -1).
+            This method returns a boolean indicating whether this position is "valid" (i.e. the index is not equal to -1).
             )mydelimiter");
 
     // py::implicitly_convertible<std::tuple<int, int>, emp::WorldPosition>();
@@ -323,17 +312,14 @@ PYBIND11_MODULE(systematics, m) {
 
         // Death notification
         .def("remove_org_by_position", static_cast<bool (sys_t::*) (emp::WorldPosition)>(&sys_t::RemoveOrg), R"mydelimiter(
-            Notify the systematics manager that an organism has died. Use this method if you're having the systematics
-            manager keep track of each organism's position and want to use the position to tell the systematics manager which
-            organism died. Otherwise use remove_org.
+            Notify the systematics manager that an organism has died. Use this method if you're having the systematics manager keep track of each organism's position and want to use the position to tell the systematics manager which organism died. Otherwise use remove_org.
 
             Parameters
             ----------
             WorldPosition position: The location of the organism that died.
             )mydelimiter")
         .def("remove_org", [](sys_t & self, taxon_t * tax){return self.RemoveOrg(tax);}, R"mydelimiter(
[…]
