+++ requirements.txt
@@ -2,7 +2,7 @@
 # This file is autogenerated by pip-compile with Python 3.9
 # by the following command:
 #
-#    pip-compile
+#    pip-compile requirements.in
 #
 alabaster==0.7.13
     # via sphinx
@@ -10,6 +10,8 @@ attrs==22.2.0
     # via pytest
 babel==2.12.1
     # via sphinx
+beautifulsoup4==4.12.2
+    # via sphinx-tippy
 certifi==2022.12.7
     # via requests
 charset-normalizer==3.1.0
@@ -25,7 +27,7 @@ idna==3.4
     # via requests
 imagesize==1.4.1
     # via sphinx
-importlib-metadata==6.0.0
+importlib-metadata==6.8.0
     # via sphinx
 iniconfig==2.0.0
     # via pytest
@@ -33,6 +35,7 @@ jinja2==3.1.2
     # via
     #   myst-parser
     #   sphinx
+    #   sphinx-tippy
 markdown-it-py==2.2.0
     # via
     #   mdit-py-plugins
@@ -60,14 +63,21 @@ pytest==7.2.2
 pyyaml==6.0
     # via myst-parser
 requests==2.28.2
-    # via sphinx
+    # via
+    #   sphinx
+    #   sphinx-tippy
 snowballstemmer==2.2.0
     # via sphinx
+soupsieve==2.4.1
+    # via beautifulsoup4
 sphinx==6.1.3
     # via
     #   myst-parser
     #   sphinx-rtd-theme
+    #   sphinx-tippy
 sphinx-rtd-theme==1.2.0
+    # via -r requirements.in
+sphinx-tippy==0.4.1
     # via -r requirements.in
 sphinxcontrib-applehelp==1.0.4
     # via sphinx
@@ -87,7 +97,7 @@ tomli==2.0.1
     # via pytest
 urllib3==1.26.14
     # via requests
-zipp==3.15.0
+zipp==3.16.2
     # via importlib-metadata
 
 # The following packages are considered to be unsafe in a requirements file:
