+++ requirements.in
@@ -2,3 +2,4 @@ pybind11
 sphinx_rtd_theme
 myst_parser
 pytest
+sphinx_tippy
