### 2023-07-18 c3d8035
Fix URLs in docstrings
