### 2023-07-18 848adca
Implement bibliography support
