+++ requirements.txt
@@ -19,8 +19,10 @@ charset-normalizer==3.1.0
 docutils==0.18.1
     # via
     #   myst-parser
+    #   pybtex-docutils
     #   sphinx
     #   sphinx-rtd-theme
+    #   sphinxcontrib-bibtex
 exceptiongroup==1.1.0
     # via pytest
 idna==3.4
@@ -28,7 +30,9 @@ idna==3.4
 imagesize==1.4.1
     # via sphinx
 importlib-metadata==6.8.0
-    # via sphinx
+    # via
+    #   sphinx
+    #   sphinxcontrib-bibtex
 iniconfig==2.0.0
     # via pytest
 jinja2==3.1.2
@@ -36,6 +40,8 @@ jinja2==3.1.2
     #   myst-parser
     #   sphinx
     #   sphinx-tippy
+latexcodec==2.0.1
+    # via pybtex
 markdown-it-py==2.2.0
     # via
     #   mdit-py-plugins
@@ -56,16 +62,28 @@ pluggy==1.0.0
     # via pytest
 pybind11==2.10.3
     # via -r requirements.in
+pybtex==0.24.0
+    # via
+    #   pybtex-docutils
+    #   sphinxcontrib-bibtex
+pybtex-docutils==1.0.2
+    # via sphinxcontrib-bibtex
 pygments==2.14.0
     # via sphinx
 pytest==7.2.2
     # via -r requirements.in
 pyyaml==6.0
-    # via myst-parser
+    # via
+    #   myst-parser
+    #   pybtex
 requests==2.28.2
     # via
     #   sphinx
     #   sphinx-tippy
+six==1.16.0
+    # via
+    #   latexcodec
+    #   pybtex
 snowballstemmer==2.2.0
     # via sphinx
 soupsieve==2.4.1
@@ -75,12 +93,15 @@ sphinx==6.1.3
     #   myst-parser
     #   sphinx-rtd-theme
     #   sphinx-tippy
+    #   sphinxcontrib-bibtex
 sphinx-rtd-theme==1.2.0
     # via -r requirements.in
 sphinx-tippy==0.4.1
     # via -r requirements.in
 sphinxcontrib-applehelp==1.0.4
     # via sphinx
+sphinxcontrib-bibtex==2.5.0
+    # via -r requirements.in
 sphinxcontrib-devhelp==1.0.2
     # via sphinx
 sphinxcontrib-htmlhelp==2.0.1
