+++ docs/bibliography.bib
@@ -1,0 +1,58 @@
+@article{tucker2017guide,
+  title={A guide to phylogenetic metrics for conservation, community ecology and macroecology},
+  author={Tucker, Caroline M and Cadotte, Marc W and Carvalho, Silvia B and Davies, T Jonathan and Ferrier, Simon and Fritz, Susanne A and Grenyer, Rich and Helmus, Matthew R and Jin, Lanna S and Mooers, Arne O and others},
+  journal={Biological Reviews},
+  volume={92},
+  number={2},
+  pages={698--715},
+  year={2017},
+  publisher={Wiley Online Library}
+}
+@article{webb2000exploring,
+  title={Exploring the phylogenetic structure of ecological communities: an example for rain forest trees},
+  author={Webb, Campbell O},
+  journal={The American Naturalist},
+  volume={156},
+  number={2},
+  pages={145--155},
+  year={2000},
+  publisher={The University of Chicago Press}
+}
+@article{clark1998artificial,
+  title={Artificial keys for botanical identification using a multilayer perceptron neural network (MLP)},
+  author={Clark, Jonathan Y and Warwick, Kevin},
+  journal={Artificial Intelligence for Biology and Agriculture},
+  pages={95--115},
+  year={1998},
+  publisher={Springer}
+}
+@article{faith1992conservation,
+  title={Conservation evaluation and phylogenetic diversity},
+  author={Faith, Daniel P},
+  journal={Biological conservation},
+  volume={61},
+  number={1},
+  pages={1--10},
+  year={1992},
+  publisher={Elsevier}
+}
+@article{winter2013phylogenetic,
+  title={Phylogenetic diversity and nature conservation: where are we?},
+  author={Winter, Marten and Devictor, Vincent and Schweiger, Oliver},
+  journal={Trends in ecology \& evolution},
+  volume={28},
+  number={4},
+  pages={199--204},
+  year={2013},
+  publisher={Elsevier}
+}
+@article{isaac2007mammals,
+  title={Mammals on the EDGE: conservation priorities based on threat and phylogeny},
+  author={Isaac, Nick JB and Turvey, Samuel T and Collen, Ben and Waterman, Carly and Baillie, Jonathan EM},
+  journal={PloS one},
+  volume={2},
+  number={3},
+  pages={e296},
+  year={2007},
+  publisher={Public Library of Science San Francisco, USA}
+}
