+++ requirements.in
@@ -3,3 +3,4 @@ sphinx_rtd_theme
 myst_parser
 pytest
 sphinx_tippy
+sphinxcontrib-bibtex

[…]
