### 2023-07-18 b424d96
Prevent sphinx from parsing API twice
This removes a ton of warnings and all duplicate indeces.
