+++ docs/conf.py
@@ -62,9 +62,9 @@ author = "Emily Dolson"
 # built documents.
 #
 # The short X.Y version.
-version = "0.1.13"
+version = "0.1.14"
 # The full version, including alpha/beta/rc tags.
-release = "0.1.13"
+release = "0.1.14"
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
