### 2023-07-23 3b8b597
Making deep copies is bad, actually
