+++ phylotrackpy/__init__.py
@@ -2,7 +2,7 @@
 
 __author__ = """Emily Dolson"""
 __email__ = 'emilydolson@gmail.com'
-__version__ = '0.1.14'
+__version__ = '0.1.15'
 
 import phylotrackpy.systematics
 
