+++ docs/conf.py
@@ -62,9 +62,9 @@ author = "Emily Dolson"
 # built documents.
 #
 # The short X.Y version.
-version = "0.1.14"
+version = "0.1.15"
 # The full version, including alpha/beta/rc tags.
-release = "0.1.14"
+release = "0.1.15"
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
