+++ systematics_bindings.cpp
@@ -78,12 +78,12 @@ PYBIND11_MODULE(systematics, m) {
     py::class_<taxon_t, std::unique_ptr<taxon_t, py::nodelete> >(m, "Taxon")
         .def(py::init<size_t, taxon_info_t>())
         .def(py::init<size_t, taxon_info_t, taxon_t*>())
-        .def("__copy__",  [](const taxon_t &self) {
+        .def("__copy__",  [](const taxon_t &self) -> const taxon_t & {
             return self;
-        })
-        .def("__deepcopy__", [](const taxon_t &self, py::dict) {
+        }, py::return_value_policy::reference_internal)
+        .def("__deepcopy__", [](const taxon_t &self, py::dict) -> const taxon_t & {
             return self;
-        })
+        }, py::return_value_policy::reference_internal)
         .def("get_parent", &taxon_t::GetParent)
         .def("get_info", &taxon_t::GetInfo)
         .def("get_id", &taxon_t::GetID, R"mydelimiter(
