+++ phylotrackpy/__init__.py
@@ -2,7 +2,7 @@
 
 __author__ = """Emily Dolson"""
 __email__ = 'emilydolson@gmail.com'
-__version__ = '0.1.15'
+__version__ = '0.1.16'
 
 import phylotrackpy.systematics
 
