+++ requirements.txt
@@ -19,13 +19,9 @@ docutils==0.18.1
     #   myst-parser
     #   sphinx
     #   sphinx-rtd-theme
-exceptiongroup==1.1.0
-    # via pytest
 idna==3.4
     # via requests
 imagesize==1.4.1
-    # via sphinx
-importlib-metadata==6.0.0
     # via sphinx
 iniconfig==2.0.0
     # via pytest
@@ -53,7 +49,7 @@ pluggy==1.0.0
     # via pytest
 pybind11==2.10.3
     # via -r requirements.in
-pygments==2.14.0
+pygments==2.15.0
     # via sphinx
 pytest==7.2.2
     # via -r requirements.in
@@ -83,12 +79,8 @@ sphinxcontrib-qthelp==1.0.3
     # via sphinx
 sphinxcontrib-serializinghtml==1.1.5
     # via sphinx
-tomli==2.0.1
-    # via pytest
 urllib3==1.26.14
     # via requests
-zipp==3.15.0
-    # via importlib-metadata
 
 # The following packages are considered to be unsafe in a requirements file:
 # setuptools
