+++ docs/bibliography.bib
@@ -56,3 +56,23 @@
   year={2007},
   publisher={Public Library of Science San Francisco, USA}
 }
+@article{mir2018sound,
+  title={Sound Colless-like balance indices for multifurcating trees},
+  author={Mir, Arnau and Rotger, Luc{\'\i}a and Rossell{\'o}, Francesc},
+  journal={PloS one},
+  volume={13},
+  number={9},
+  pages={e0203401},
+  year={2018},
+  publisher={Public Library of Science San Francisco, CA USA}
+}
+@article{shao1990tree,
+  title={Tree balance},
+  author={Shao, Kwang-Tsao and Sokal, Robert R},
+  journal={Systematic Zoology},
+  volume={39},
+  number={3},
+  pages={266--276},
+  year={1990},
+  publisher={Society of Systematic Zoology}
+}
