### 2023-07-27 #5660 @Kevin-Mattheus-Moerman
@emilydolson thanks for this submission. I am the AEiC for this track and here to help with initial steps. Note that I've just flagged this work for a scope review by the editorial board. This is because of the apparent relatively small size of this submission in terms of LOC. Furthermore the limited functionality provided, as this work appears to be "essentially a wrapper around Phylotracklib" as your README states. Please let us know if the above line counts misrepresent the magnitude of the project of if some aspects are missing from this analysis.

The editorial board will now review this submission to see if it conforms to our [substantial scholarly effort criteria](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). This review should take around 2 weeks to complete.
