### 2023-07-31 #5660 @arfon
> They are split across two repositories because phylotracklib is a module in a larger piece of scientific software (included as a submodule in phylotrackpy)

@emilydolson – could you link to that code so we can make a judgment on scope that includes that contribution too?

Also, apologies for the lack of response on a pre-submission enquiry. You did the right thing just submitting anyway :-)
