### 2023-07-31 #5660 @emilydolson
Oh yes, sorry, here it is: https://github.com/devosoft/Empirical/blob/master/include/emp/Evolve/Systematics.hpp. The SystematicsAnalysis.hpp file in the same directory is also a smaller part of it.
