### 2023-08-01 #5660 @Kevin-Mattheus-Moerman
@emilydolson I have just removed the query-scope label as with the above additional repo link the total magnitude of this work seems like it may be in scope for JOSS. Hence I will now proceed to find a handling editor. Note though that many are currently somewhat oversubscribed or on holiday. So it may take some time before we find one, and I may need to waitlist this submission if none are available at the moment.
