### 2023-08-07 #5660 @emilydolson
I think @/CFGrote, @/mtomasini and @/stulacy could be good choices from the JOSS reviewer list! Some good choices not from the JOSS reviewer list might be @/RussellGarwood (since I just reviewed a JOSS paper from him that involved some related ideas) or @/Tim-Taylor.
