### 2023-08-09 #5660 @RussellGarwood
@fboehm I'd be very happy to review this contribution (I note that some of my experience is well aligned with reviewing this, but there are some areas this touches on in which I'm not so confident. Presumably I can just highlight these if required).
