### 2023-08-10 #5660 @CFGrote
> > @CFGrote @mtomasini@stulacy @RussellGarwood @tim-taylor - would you like to review this submission for JOSS? Unlike those at many journals, our reviews are checklist-driven and involve discussions with the authors. Thank you for considering this invitation!
>
> @fboehm I'm sorry but I don't have time to devote to this at present so I will have to decline. Apologies. Just a comment about the invitation - it would be helpful if you could more clearly state exactly what you are requesting is reviewed (the paper, the code, or both) and provide links to these in the invitation. Best wishes, Tim

@tim-taylor, I had the same initial thoughts, but then realized that following the link to this issue provides all the information you requested.
