### 2023-08-10 #5660 @fboehm
@tim-taylor and @CFGrote - thank you for the comments on the text of the invitation. I'll try to ensure that the relevant links are included in the text of the invitation message itself in the future. Thanks again!
