### 2023-08-10 #5754 @editorialbot
[REVIEW]: Phylotrack: C++ and Python libraries for in silico phylogenetic tracking
**Submitting author:** @emilydolson (<a href="http://orcid.org/0000-0001-8616-4898">Emily Dolson</a>)
**Repository:** https://github.com/emilydolson/phylotrackpy
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.13
