### 2023-08-11 emilydolson/python-phylogeny-tracker#32 @CFGrote
JOSS Review: Installation
[JOSS Review issue](https://github.com/openjournals/joss-reviews/issues/5754)

Installation proceeds as claimed in the README: `pip install phylotrack`. However, I would suggest adding instructions how to
install the latest development snapshot (e.g. via `pip git+https://github.com/emilydolson/phylotrackpy`) and how to install from sources. The latter is important for developers and potential contributors. Could also link to the corresponding section in phylotrackpylotrackpy/CONTRIBUTING.md .
