### 2023-08-11 emilydolson/python-phylogeny-tracker#33 @CFGrote
JOSS Review: Documentation hosting and building
[JOSS Review issue](https://github.com/openjournals/joss-reviews/issues/5754)

I'd suggest publishing the documentation e.g. on readthedocs.org or as gh pages. Lacking that, I build the documentation locally in the cloned repo and came across/observed several (potential) issues

- [ ] Users/contributors could benefit from instructions how to install dependencies (code and docs)
- [ ] Users/contributors could benefit from instructions how to build the documentation
- [ ] On a Linux system, the env. variable LC_ALL must be set, otherwise `make html` in docs/ errors out.
- [ ] Packages sphinx_tippy and sphinxcontrib_bibtex should be added to doc dependencies
