### 2023-08-11 emilydolson/python-phylogeny-tracker#34 @CFGrote
JOSS Review: Functionality, documentation
[JOSS Review issue](https://github.com/openjournals/joss-reviews/issues/5754)

# General documentation
I very much appreciate the detailled introduction and explanations regarding the subtleties in running in-silico vs. in-vitro or in-vivo phylogenetic tracking.

# Usage example
 That being said, I'm missing similar in-depth usage instructions at a more concrete level than what is
provided in the docs. The usage example does not go beyond showing how to instantiate the `Systematics` class. What can I do with the class from there on? I tried adding an organism, but needed to mock up an `Organism` class first. Unfortunately, I'm left
entirely on my own with that task. Would be helpful to have some minimum requirements for such an object or even better a template (abstract base class).

# Writing an Organism class
 Here's my attempt:

```
from phylotrackpy import systematics

syst = systematics.Systematics(lambda org: org.genotype)

class Organism():
    genotype = "ACTG"

org0 = Organism()

org0.genotype

syst.add_org(org0)
```

This code lived in a jupyter notebook, executing the last line of code crashed the kernel.

# Coding style
I would recommend not to use `sys` for a local variable name as it obscures the python native library of the same name (which many python codes import)

# Usage example recommendation
For a proper usage example I would expect to be instructed how to build a phylogeny from scratch and how to analyse it. You could also provide the data for below figure (from the README) and instruct the user how to reproduce it.

![Copied from README.md](https://github.com/emilydolson/phylotrackpy/raw/main/docs/images/FullPhylogeny.png)

# API documentation
Please complete the API documentation and, if done, remove the note "Partially under construction". Probably related to #13 and #22
