### 2023-08-11 emilydolson/python-phylogeny-tracker#35 @CFGrote
JOSS Review: Paper
[JOSS Review issue](https://github.com/openjournals/joss-reviews/issues/5754)

The paper reads great and leaves only few minor details to be desired:

- line 164: Ref. "Moreno et al (under review)" seems to be unpublished, is it possible to link to a preprint instead?
- There is no URL link for the reference to pybind (Jakob, 2017).
