### 2023-08-11 emilydolson/python-phylogeny-tracker#36 @CFGrote
JOSS Review: Input data format conversion
[JOSS Review issue](https://github.com/openjournals/joss-reviews/issues/5754)

This is not a critical review issue but I came across it while reviewing:

Reading phylogenies stored in foreign file formats seems not to be supported.

**Describe the solution you'd like**
I'm running stochastic simulations and store the resulting trajectories in a newick tree file or in a json document. Now I want to analyse my data with phylotracks. How can I get the data into phylotracks?
