### 2023-08-18 emilydolson/python-phylogeny-tracker#33 @emilydolson
Thanks so much for all of your feedback! I've been traveling, but I'm excited to get it all implemented because I think it will really improve usability for new users.

We actually do have the documentation on readthedocs (https://phylotrackpy.readthedocs.io/en/latest/?badge=latest) but clearly the link is not as prominent as it should be! We'll make it clearer, and add the instructions for building locally that you suggest.

One quick question: which linux distribution are you using? I'm on PopOS and do not need to set LC_ALL, so I'm assuming it's a distro-specific thing and want to mention that in the docs.
