### 2023-08-21 emilydolson/python-phylogeny-tracker#33 @CFGrote
My linux distro is ubuntu 22.04. I have to admit that I have some uncommon language and region settings (combination of US keyboard and German regional/language settings) which may have triggered the LC_ALL settings issue. But maybe worth checking on a vanilla ubuntu installation. github actions is your friend ;)
