### 2023-08-29 emilydolson/python-phylogeny-tracker#34 @RussellGarwood
[JOSS Review issue](https://github.com/openjournals/joss-reviews/issues/5754)

I am also reviewing this for JOSS, and one of my requests was going to be a usage example - so I thought I would just note here that I concur, this would be a very valuable addition!
