### 2023-08-29 emilydolson/python-phylogeny-tracker#35 @RussellGarwood
[JOSS Review issue](https://github.com/openjournals/joss-reviews/issues/5754)

I note the paper states the software can "support large agent populations with rapid generational turnover" - it might be useful in this section to indicate ballpark figures for large, and rapid, within this context, as I imagine this differs a good deal between models.
