### 2023-08-29 emilydolson/python-phylogeny-tracker#37 @RussellGarwood
JOSS review - exporting data
[JOSS Review issue](https://github.com/openjournals/joss-reviews/issues/5754)

Often it is likely we'll want to do a downstream analysis on evolutionary trees from a simulatrion - for instance, beyond the output available within Phylotrack, I may want to interrogate originations and extinctions of my taxa in another language, and conduct some comparative phylogenetic methods on that in concert with my genotypes.

**Describe the solution you'd like**
From interrogating the docs and the paper, I see that the data can be output in a  Artificial Life Phylogeny Data Standard format, and the paper suggests that can be converted to e.g. Newick. When thinking about adding usage examples, to your documentation (issue #34), would it be possible to add a workflow that demonstrates this for the uninitiated? I think that emphasising the intraoperability of this tool will highlight its strengths and potential uses!
