### 2023-10-24 466c289
Add first-time resources to CONTRIBUTING
