+++ docs/community.md
@@ -1,0 +1,29 @@
+# Project Community
+
+## Communications
+
+Join the Phylotrack [mailing list](https://groups.google.com/g/phylotrack) for project announcements and discussion.
+If you'd like to keep up with day-to-day project chat, contact a maintainer to be added to our Slack workspace.
+
+## Projects Using PhylotrackLib
+
+Open [an issue](https://github.com/emilydolson/phylotrackpy/issues) or [a pull request](https://github.com/emilydolson/phylotrackpy/compare) to list your project here!
+
+## Contributors
+
+- Emily Dolson [@emilydolson](https://github.com/emilydolson) (maintainer)
+- Matthew Andres Moreno [@mmore500](https://github.com/mmore500) (maintainer)
+- Santiago Rodriguez Papa [@rodsan0](https://github.com/rodsan0)
+
+Find a comprehensive, up-to-date listing of project contributors [here](https://github.com/emilydolson/phylotrackpy/graphs/contributors).
+
+## Governance
+
+As an early-stage project, Phylotrack currently operates under a founder-leader model, where project decisions are driven by consensus within a small group of core maintainers.
+[@emilydolson](https://github.com/emilydolson) and [@mmore500](https://github.com/mmore500) currently lead in this role.
+
+We are committed to actively involving Phylotrack’s user community in project decision-making.
+To weigh in on project directions, please open feature request(s) through our [GitHub issue tracker](https://github.com/emilydolson/phylotrackpy/issues) and Be notified of announcements and join in on discussions through mailing list, which you can join [here](https://groups.google.com/g/phylotrack).
+
+```{include} code_of_conduct.md
+```
