### 2023-11-02 #5754 @emilydolson
Thanks for checking in - I'm okay, just very overwhelmed. No questions yet and I will definitely let you know if I have any. Really hoping to finish this soon, it's just been slow going.
