+++ requirements.txt
@@ -84,7 +84,6 @@ six==1.16.0
     # via
     #   latexcodec
     #   pybtex
-    # via myst-parser
 snowballstemmer==2.2.0
     # via sphinx
 soupsieve==2.4.1
@@ -115,7 +114,9 @@ sphinxcontrib-qthelp==1.0.3
     # via sphinx
 sphinxcontrib-serializinghtml==1.1.5
     # via sphinx
-urllib3==1.26.14
+tomli==2.0.1
+    # via pytest
+urllib3==1.26.18
     # via requests
 zipp==3.16.2
     # via importlib-metadata
