+++ docs/conf.py
@@ -68,9 +68,9 @@ author = "Emily Dolson"
 # built documents.
 #
 # The short X.Y version.
-version = "0.1.16"
+version = "0.1.17"
 # The full version, including alpha/beta/rc tags.
-release = "0.1.16"
+release = "0.1.17"
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
