+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.1.16
+current_version = 0.1.17
 commit = True
 tag = True
 
@@ -18,3 +18,4 @@ replace = __version__ = '{new_version}'
 [metadata]
 long_description = file: README.md
 long_description_content_type = text/markdown
+
