### 2023-11-22 emilydolson/python-phylogeny-tracker#36 @emilydolson
Thanks so much for this suggestion! @mmore500 has added support for this via his alifedata-phyloinformatics-convert library, which is designed to handle translation among the format used by phylotrackpy and more standard bioinformatics formats. We have added documentation to the quickstart guide explaining how to use it. I'm going to close this for now, but if you see ways the documentation on it could be improved let me know!
