### 2023-11-22 emilydolson/python-phylogeny-tracker#37 @emilydolson
Great suggestion! We have added such a workflow to the documentation. I'm going to close this for now, but let me know if you think what we added is unclear.
