+++ profile/profile_batch.sh
@@ -1,0 +1,211 @@
+#!/bin/bash
+
+BATCH_UUID="$(uuid)"
+
+{
+
+###############################################################################
+echo "SETUP"
+echo "########################################################################"
+cd "$(dirname "$0")"
+echo "pwd $(pwd)"
+
+RUN_SECONDS=4
+echo "RUN_SECONDS ${RUN_SECONDS}"
+
+echo "BATCH_UUID ${BATCH_UUID}"
+mkdir -p "batch=${BATCH_UUID}+what=rawdata"
+
+VENV="$(mktemp -d)"
+python3 -m venv "${VENV}"
+source "${VENV}/bin/activate"
+python3 -m pip install --upgrade pip
+python3 -m pip install -r requirements.txt
+
+
+###############################################################################
+echo "MAKE PROVLOG"
+echo "########################################################################"
+  cat << EOF >> "batch=${BATCH_UUID}+what=provlog+ext=.yaml"
+-
+  a: "batch=${BATCH_UUID}+what=provlog+ext=.yaml"
+  batch: ${BATCH_UUID}
+  date: \$(date --iso-8601=seconds)
+  python: $(python3 --version)
+  revision: $(git rev-parse HEAD)
+  hostname: $(hostname)
+  pwd: $(pwd)
+  neofetch:
+$(neofetch --stdout | sed 's/^/    /')
+  os_info: |
+$(lsb_release -a 2>/dev/null || cat /etc/*release | sed 's/^/    /')
+  cpu_info: |
+$(lscpu | sed 's/^/    /')
+  memory_info: |
+$(free -h | sed 's/^/    /')
+  disk_info: |
+$(df -h | sed 's/^/    /')
+  network_info: |
+$(ifconfig -a | sed 's/^/    /')
+  user: $(whoami)
+  uuid: $(uuidgen)
+  slurm_job_id: ${SLURM_JOB_ID-none}
+  profile_batch.sh: |
+$(sed 's/^/    /' profile_batch.sh)
+  profile_job.py: |
+$(sed 's/^/    /' profile_job.py)
+  requirements.txt: |
+$(sed 's/^/    /' requirements.txt)
+EOF
+
+###############################################################################
+echo "RUN JOBS"
+echo "########################################################################"
+for replicate in {0..4}; do for pop_size in 10 1000 100000; do
+    JOB_UUID="$(uuid)"
+    echo "JOB_UUID ${JOB_UUID} replicate ${replicate} pop_size ${pop_size}"
+    mprof run -o "batch=${BATCH_UUID}+what=rawdata/a=memprof+job=${JOB_UUID}+ext=.dat" \
+        python3 profile_job.py \
+        "${BATCH_UUID}" "${JOB_UUID}" \
+        "${replicate}" "${pop_size}"
+done; done
+
+
+###############################################################################
+echo "COLLLATE DATA"
+echo "########################################################################"
+python3 << EOF
+import glob
+
+from keyname import keyname as kn
+import pandas as pd
+
+
+# COLLATE TIME DATA
+###############################################################################
+time_file_list = glob.glob("batch=${BATCH_UUID}+what=rawdata/a=timeprof+job=*+ext=.csv")
+time_concat = pd.concat([pd.read_csv(f) for f in time_file_list])
+time_concat.to_csv("batch=${BATCH_UUID}+what=timeprof+ext=.csv", index=False)
+
+# COLLATE MEMORY DATA
+##############################################################################
+mem_file_list = glob.glob("batch=${BATCH_UUID}+what=rawdata/a=memprof+job=*+ext=.dat")
+
+# Read and concatenate the CSV files
+mem_data = []
+for f in mem_file_list:
+    job = kn.unpack(f)["job"]
+    mem_df = pd.read_csv(
[…]
