### 2023-11-22 e7f953e
Plot evaluations per second, too
