### 2023-11-22 28d842e
Delete unused variable
