### 2023-11-22 b619f37
Clear plot between draws
