+++ joss/paper.bib
@@ -151,6 +151,12 @@
   year = 2017,
   note = {https://github.com/pybind/pybind11},
 }
+@misc{memory_profiler,
+  title = {{Memory Profiler}},
+  author = {Fabian Pedregosa and Philippe Gervais},
+  year = 2023,
+  note = {https://github.com/pythonprofilers/memory_profiler},
+}
 @inproceedings{shahbandegan2022untangling,
   title = {Untangling Phylogenetic Diversity's Role in Evolutionary Computation Using a Suite of Diagnostic Fitness Landscapes},
   author = {Shahbandegan, Shakiba and Hernandez, Jose Guadalupe and Lalejini, Alexander and Dolson, Emily},
@@ -552,4 +558,16 @@ would give a high Λ+ by comparison with a list (of equivalent Δ+) in which all
   title = {Lineage Tracking Algorithms and the Streaming Curation Problem},
   author = {Moreno, Matthew Andres and Rodriguez-Papa, Santiago and Dolson, Emily},
   year = {in review}
+}
+@article{foster2017open,
+  title     = {Open science framework (OSF)},
+  author    = {Foster, Erin D and Deardorff, Ariel},
+  journal   = {Journal of the Medical Library Association: JMLA},
+  volume    = {105},
+  number    = {2},
+  pages     = {203},
+  year      = {2017},
+  doi       = {10.5195/jmla.2017.88},
+  url={https://doi.org/10.5195/jmla.2017.88},
+  publisher = {Medical Library Association}
 }
