### 2023-11-23 92ab54a
Drop unnecessary lambda
