### 2023-11-23 57e3859
Implement, test py::object as taxon type
Probably could get a speed boost by specializing for string, float, int,
etc. instead of using py::object but this works!
