+++ setup.py
@@ -40,6 +40,7 @@ setup(
     zip_safe=False,
     python_requires=">=3.6",
     install_requires=[
+        "numpy",
     ],
     packages=['phylotrackpy'],
 )
