+++ requirements.txt
@@ -54,6 +54,8 @@ mdurl==0.1.2
     # via markdown-it-py
 myst-parser==0.19.1
     # via -r requirements.in
+numpy==1.26.2
+    # via -r requirements.in
 packaging==23.0
     # via
     #   pytest
@@ -72,8 +74,6 @@ pygments==2.15.0
     # via sphinx
 pytest==7.2.2
     # via -r requirements.in
-pytz==2023.3.post1
-    # via babel
 pyyaml==6.0.1
     # via
     #   -r requirements.in
