+++ requirements.in
@@ -4,6 +4,7 @@ myst_parser
 pytest
 sphinx_tippy
 sphinxcontrib-bibtex
+numpy
 
 # need pin for python3.12 compatibility
 # see https://github.com/yaml/pyyaml/issues/723
