### 2023-11-23 f50db23
Use oldest-supported-numpy for cross-release pin
