+++ requirements.txt
@@ -2,7 +2,7 @@
 # This file is autogenerated by pip-compile with Python 3.9
 # by the following command:
 #
-#    pip-compile requirements.in
+#    pip-compile --unsafe-package=numpy --unsafe-package=setuptools requirements.in
 #
 alabaster==0.7.13
     # via sphinx
@@ -23,16 +23,10 @@ docutils==0.18.1
     #   sphinx
     #   sphinx-rtd-theme
     #   sphinxcontrib-bibtex
-exceptiongroup==1.1.0
-    # via pytest
 idna==3.4
     # via requests
 imagesize==1.4.1
     # via sphinx
-importlib-metadata==6.8.0
-    # via
-    #   sphinx
-    #   sphinxcontrib-bibtex
 iniconfig==2.0.0
     # via pytest
 jinja2==3.1.2
@@ -54,7 +48,7 @@ mdurl==0.1.2
     # via markdown-it-py
 myst-parser==0.19.1
     # via -r requirements.in
-numpy==1.26.2
+oldest-supported-numpy==0.5
     # via -r requirements.in
 packaging==23.0
     # via
@@ -117,12 +111,9 @@ sphinxcontrib-qthelp==1.0.3
     # via sphinx
 sphinxcontrib-serializinghtml==1.1.5
     # via sphinx
-tomli==2.0.1
-    # via pytest
 urllib3==1.26.18
     # via requests
-zipp==3.16.2
-    # via importlib-metadata
 
 # The following packages are considered to be unsafe in a requirements file:
+# numpy
 # setuptools
