### 2023-11-23 848962a
Replace std::format use
Because <format> isn't available on compilers in CI
