### 2023-11-23 6d3fbeb
Bugfix: support optional desc in add_snapshot_fun
