### 2023-11-23 b70b3dd
Fixup, test full taxon info save/load
TODO: Empirical should support quote escaping csv entries containing
"," so that we don't have to url encode taxon info representations
containg ","
