### 2023-11-23 468818f
Document and iron env setup and builds
