### 2023-11-23 5b3c440
Document and iron env setup and builds
