+++ docs/Makefile
@@ -217,3 +217,7 @@ pseudoxml:
 	$(SPHINXBUILD) -b pseudoxml $(ALLSPHINXOPTS) $(BUILDDIR)/pseudoxml
 	@echo
 	@echo "Build finished. The pseudo-XML files are in $(BUILDDIR)/pseudoxml."
+
+.PHONY: serve
+serve:
+	cd $(BUILDDIR)/html && python -m http.server
