### 2023-11-23 35d2d8a
Add instructions for github, local installs
