### 2023-11-24 6a17df1
Line edit and trim profiling section
