### 2023-11-24 ff2e1ce
Make line edits to condense
