+++ test/test_systematics.py
@@ -1,8 +1,16 @@
+#!/usr/bin/env python3
 from phylotrackpy import systematics
 from pytest import approx, mark
 from copy import deepcopy
 import numpy as np
 import tempfile
+
+
+# wheel build CI components call this file directly
+if __name__ == "__main__":
+    import pytest
+
+    pytest.main()
 
 
 class ExampleOrg:
