### 2023-11-24 a6611ca
Try pinning windows CI to oldest-supported-numpy
