### 2023-11-24 748f5f4
Use older oldest-supported-numpy
Apparently new release isn't compat with old windows python
