### 2023-11-24 997e78f
Add pytest dep to windows build
