### 2023-11-24 21e9184
Pin windows to older pytest
