### 2023-11-24 3945567
Provide explicit test command to CI build wheel
