### 2023-11-24 1034000
Specify asset path independent of pwd
