### 2023-11-24 a078bb4
Use {project} template instead of hardcoding
