### 2023-11-24 ec762cf
Try excluding python3.6 build
