### 2023-11-24 b3bed86
Make build sdist run right away, too
