### 2023-11-24 281f4e5
Test build sdist in non-deployment CI
