+++ .github/workflows/ci.yml
@@ -75,7 +75,6 @@ jobs:
     name: Build source distribution
     needs: []
     runs-on: ubuntu-latest
-    if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags/v')
     steps:
       - uses: actions/checkout@v3
         with:
