### 2023-11-24 b1871ec
Restore py36, try excluding local source discovery
