### 2023-11-24 d4d76c5
Ignore python version requirement in test install
