+++ .github/workflows/ci.yml
@@ -31,7 +31,8 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install flake8 pytest pytest-cov
-        if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+        # --ignore-requires-python ensures docs deps don't break install on early python versions
+        if [ -f requirements.txt ]; then pip install --ignore-requires-python -r requirements.txt; fi
         pip install -e .
     - name: Lint with flake8
       run: |
