### 2023-11-24 aa56db1
Move flake8, pytest-cov to requirements
