+++ requirements.txt
@@ -16,6 +16,8 @@ certifi==2023.7.22
     # via requests
 charset-normalizer==3.1.0
     # via requests
+coverage[toml]==7.3.2
+    # via pytest-cov
 docutils==0.18.1
     # via
     #   myst-parser
@@ -23,10 +25,19 @@ docutils==0.18.1
     #   sphinx
     #   sphinx-rtd-theme
     #   sphinxcontrib-bibtex
+exceptiongroup==1.2.0
+    # via pytest
+flake8==6.1.0
+    # via -r requirements.in
 idna==3.4
     # via requests
 imagesize==1.4.1
     # via sphinx
+importlib-metadata==5.2.0
+    # via
+    #   -r requirements.in
+    #   sphinx
+    #   sphinxcontrib-bibtex
 iniconfig==2.0.0
     # via pytest
 jinja2==3.1.2
@@ -42,6 +53,8 @@ markdown-it-py==2.2.0
     #   myst-parser
 markupsafe==2.1.2
     # via jinja2
+mccabe==0.7.0
+    # via flake8
 mdit-py-plugins==0.3.5
     # via myst-parser
 mdurl==0.1.2
@@ -64,10 +77,20 @@ pybtex==0.24.0
     #   sphinxcontrib-bibtex
 pybtex-docutils==1.0.2
     # via sphinxcontrib-bibtex
+pycodestyle==2.11.1
+    # via flake8
+pyflakes==3.1.0
+    # via flake8
 pygments==2.15.0
     # via sphinx
 pytest==7.2.2
+    # via
+    #   -r requirements.in
+    #   pytest-cov
+pytest-cov==4.1.0
     # via -r requirements.in
+pytz==2023.3.post1
+    # via babel
 pyyaml==6.0.1
     # via
     #   -r requirements.in
@@ -111,8 +134,14 @@ sphinxcontrib-qthelp==1.0.3
     # via sphinx
 sphinxcontrib-serializinghtml==1.1.5
     # via sphinx
+tomli==2.0.1
+    # via
+    #   coverage
+    #   pytest
 urllib3==1.26.18
     # via requests
+zipp==3.17.0
+    # via importlib-metadata
 
 # The following packages are considered to be unsafe in a requirements file:
 # numpy
