+++ requirements.in
@@ -1,11 +1,14 @@
 pybind11
 sphinx_rtd_theme
+pytest_cov
 myst_parser
 pytest
 sphinx_tippy
 sphinxcontrib-bibtex
 oldest-supported-numpy
+flake8
 
 # need pin for python3.12 compatibility
 # see https://github.com/yaml/pyyaml/issues/723
 pyyaml>=6.0.1
+importlib_metadata<6  # for py36 compat
