### 2023-11-24 db3556b
Pin older flake8 for py36/7 compat
