### 2023-11-25 1c4df10
Pin zipp for py36 compat
