+++ requirements.in
@@ -12,3 +12,4 @@ flake8<5
 # see https://github.com/yaml/pyyaml/issues/723
 pyyaml>=6.0.1
 importlib_metadata<6  # for py36 compat
+zipp<3  # for py36 compat
