+++ requirements.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-python3 -m piptools compile requirements.in --unsafe-package numpy --unsafe-package setuptools
+python3 -m piptools compile requirements.in --unsafe-package numpy --unsafe-package setuptools -c constraints.in
