+++ requirements.txt
@@ -35,7 +35,7 @@ imagesize==1.4.1
     # via sphinx
 importlib-metadata==5.2.0
     # via
-    #   -r requirements.in
+    #   -c constraints.in
     #   sphinx
     #   sphinxcontrib-bibtex
 iniconfig==2.0.0
@@ -93,7 +93,7 @@ pytz==2023.3.post1
     # via babel
 pyyaml==6.0.1
     # via
-    #   -r requirements.in
+    #   -c constraints.in
     #   myst-parser
     #   pybtex
 requests==2.31.0
@@ -142,7 +142,7 @@ urllib3==1.26.18
     # via requests
 zipp==2.2.1
     # via
-    #   -r requirements.in
+    #   -c constraints.in
     #   importlib-metadata
 
 # The following packages are considered to be unsafe in a requirements file:
