+++ requirements.in
@@ -7,9 +7,3 @@ sphinx_tippy
 sphinxcontrib-bibtex
 oldest-supported-numpy
 flake8<5
-
-# need pin for python3.12 compatibility
-# see https://github.com/yaml/pyyaml/issues/723
-pyyaml>=6.0.1
-importlib_metadata<6  # for py36 compat
-zipp<3  # for py36 compat
