+++ requirements.in
@@ -7,3 +7,5 @@ sphinx_tippy
 sphinxcontrib-bibtex
 oldest-supported-numpy
 flake8<5
+
+typing-extensions<3.8  # for py36 compat
