### 2023-11-25 429f9bb
Pin coverage pkg for py36 compat
