+++ requirements.txt
@@ -16,7 +16,7 @@ certifi==2023.7.22
     # via requests
 charset-normalizer==3.1.0
     # via requests
-coverage[toml]==6.2
+coverage[toml]==6.1.2
     # via
     #   -c constraints.in
     #   pytest-cov
