+++ requirements.txt
@@ -16,7 +16,7 @@ certifi==2023.7.22
     # via requests
 charset-normalizer==3.1.0
     # via requests
-coverage[toml]==6.1.2
+coverage[toml]==5.5
     # via
     #   -c constraints.in
     #   pytest-cov
@@ -27,19 +27,12 @@ docutils==0.18.1
     #   sphinx
     #   sphinx-rtd-theme
     #   sphinxcontrib-bibtex
-exceptiongroup==1.2.0
-    # via pytest
 flake8==2.3.0
     # via -r requirements.in
 idna==3.4
     # via requests
 imagesize==1.4.1
     # via sphinx
-importlib-metadata==5.2.0
-    # via
-    #   -c constraints.in
-    #   sphinx
-    #   sphinxcontrib-bibtex
 iniconfig==2.0.0
     # via pytest
 jinja2==3.1.2
@@ -91,8 +84,6 @@ pytest==7.2.2
     #   pytest-cov
 pytest-cov==4.1.0
     # via -r requirements.in
-pytz==2023.3.post1
-    # via babel
 pyyaml==6.0.1
     # via
     #   -c constraints.in
@@ -136,18 +127,12 @@ sphinxcontrib-qthelp==1.0.3
     # via sphinx
 sphinxcontrib-serializinghtml==1.1.5
     # via sphinx
-tomli==2.0.1
-    # via
-    #   coverage
-    #   pytest
+toml==0.10.2
+    # via coverage
 typing-extensions==3.7.4.3
     # via -r requirements.in
 urllib3==1.26.18
     # via requests
-zipp==2.2.1
-    # via
-    #   -c constraints.in
-    #   importlib-metadata
 
 # The following packages are considered to be unsafe in a requirements file:
 # numpy
