+++ constraints.in
@@ -3,4 +3,4 @@
 pyyaml>=6.0.1
 importlib_metadata<6  # for py36 compat
 zipp<3  # for py36 compat
-coverage<6.2
+coverage<6
