### 2023-11-25 3ae72b0
Add python3.6 to ci test matrix
