+++ requirements.txt
@@ -33,6 +33,10 @@ idna==3.4
     # via requests
 imagesize==1.4.1
     # via sphinx
+importlib-metadata==5.2.0
+    # via
+    #   -c constraints.in
+    #   -r requirements.in
 iniconfig==2.0.0
     # via pytest
 jinja2==3.1.2
@@ -133,6 +137,11 @@ typing-extensions==3.7.4.3
     # via -r requirements.in
 urllib3==1.26.18
     # via requests
+zipp==2.2.1
+    # via
+    #   -c constraints.in
+    #   -r requirements.in
+    #   importlib-metadata
 
 # The following packages are considered to be unsafe in a requirements file:
 # numpy
