+++ requirements.in
@@ -9,3 +9,5 @@ oldest-supported-numpy
 flake8<5
 
 typing-extensions<3.8  # for py36 compat
+zipp<3  # for py36 compat
+importlib_metadata<6  # for py36 compat
