### 2023-11-25 ed7f9b8
Drop python3.6 support and testing
Py36 binaries aren't available on gh actions, so can't test it.
