### 2023-11-26 8634e7b
Try adding asserts
