+++ setup.py
@@ -39,8 +39,5 @@ setup(
     cmdclass={"build_ext": build_ext},
     zip_safe=False,
     python_requires=">=3.7",
-    install_requires=[
-        "numpy",
-    ],
     packages=['phylotrackpy'],
 )
