+++ requirements.txt
@@ -2,7 +2,7 @@
 # This file is autogenerated by pip-compile with Python 3.9
 # by the following command:
 #
-#    pip-compile --unsafe-package=numpy --unsafe-package=setuptools requirements.in
+#    pip-compile
 #
 alabaster==0.7.13
     # via sphinx
@@ -17,9 +17,7 @@ certifi==2023.7.22
 charset-normalizer==3.1.0
     # via requests
 coverage[toml]==5.5
-    # via
-    #   -c constraints.in
-    #   pytest-cov
+    # via pytest-cov
 docutils==0.18.1
     # via
     #   myst-parser
@@ -27,6 +25,8 @@ docutils==0.18.1
     #   sphinx
     #   sphinx-rtd-theme
     #   sphinxcontrib-bibtex
+exceptiongroup==1.2.0
+    # via pytest
 flake8==2.3.0
     # via -r requirements.in
 idna==3.4
@@ -35,8 +35,9 @@ imagesize==1.4.1
     # via sphinx
 importlib-metadata==5.2.0
     # via
-    #   -c constraints.in
     #   -r requirements.in
+    #   sphinx
+    #   sphinxcontrib-bibtex
 iniconfig==2.0.0
     # via pytest
 jinja2==3.1.2
@@ -59,8 +60,6 @@ mdit-py-plugins==0.3.5
 mdurl==0.1.2
     # via markdown-it-py
 myst-parser==0.19.1
-    # via -r requirements.in
-oldest-supported-numpy==0.5
     # via -r requirements.in
 packaging==23.0
     # via
@@ -90,7 +89,6 @@ pytest-cov==4.1.0
     # via -r requirements.in
 pyyaml==6.0.1
     # via
-    #   -c constraints.in
     #   myst-parser
     #   pybtex
 requests==2.31.0
@@ -133,16 +131,16 @@ sphinxcontrib-serializinghtml==1.1.5
     # via sphinx
 toml==0.10.2
     # via coverage
+tomli==2.0.1
+    # via pytest
 typing-extensions==3.7.4.3
     # via -r requirements.in
 urllib3==1.26.18
     # via requests
 zipp==2.2.1
     # via
-    #   -c constraints.in
     #   -r requirements.in
     #   importlib-metadata
 
 # The following packages are considered to be unsafe in a requirements file:
-# numpy
 # setuptools
