### 2023-11-26 9177964
Make test command more general
