+++ pyproject.toml
@@ -8,7 +8,7 @@ requires = [
 build-backend = "setuptools.build_meta"
 
 [tool.cibuildwheel]
-test-command = "pytest {project}/test/test_systematics.py"
+test-command = "pytest {project}"
 skip = "*musllinux*"
 test-requires = "pytest"
 before-test = "cp {project}/test/assets/* ."
