+++ requirements.txt
@@ -2,22 +2,22 @@
 # This file is autogenerated by pip-compile with Python 3.9
 # by the following command:
 #
-#    pip-compile
+#    pip-compile --resolver=backtracking
 #
 alabaster==0.7.13
     # via sphinx
-attrs==22.2.0
-    # via pytest
-babel==2.12.1
+babel==2.13.1
     # via sphinx
 beautifulsoup4==4.12.2
     # via sphinx-tippy
-certifi==2023.7.22
+certifi==2023.11.17
     # via requests
-charset-normalizer==3.1.0
+charset-normalizer==3.3.2
     # via requests
-coverage[toml]==5.5
-    # via pytest-cov
+coverage[toml]==7.3.2
+    # via
+    #   coverage
+    #   pytest-cov
 docutils==0.18.1
     # via
     #   myst-parser
@@ -27,15 +27,14 @@ docutils==0.18.1
     #   sphinxcontrib-bibtex
 exceptiongroup==1.2.0
     # via pytest
-flake8==2.3.0
+flake8==6.1.0
     # via -r requirements.in
-idna==3.4
+idna==3.6
     # via requests
 imagesize==1.4.1
     # via sphinx
-importlib-metadata==5.2.0
+importlib-metadata==6.8.0
     # via
-    #   -r requirements.in
     #   sphinx
     #   sphinxcontrib-bibtex
 iniconfig==2.0.0
@@ -47,41 +46,41 @@ jinja2==3.1.2
     #   sphinx-tippy
 latexcodec==2.0.1
     # via pybtex
-markdown-it-py==2.2.0
+markdown-it-py==3.0.0
     # via
     #   mdit-py-plugins
     #   myst-parser
-markupsafe==2.1.2
+markupsafe==2.1.3
     # via jinja2
 mccabe==0.7.0
     # via flake8
-mdit-py-plugins==0.3.5
+mdit-py-plugins==0.4.0
     # via myst-parser
 mdurl==0.1.2
     # via markdown-it-py
-myst-parser==0.19.1
+myst-parser==2.0.0
     # via -r requirements.in
-packaging==23.0
+packaging==23.2
     # via
     #   pytest
     #   sphinx
-pep8==1.7.1
-    # via flake8
-pluggy==1.0.0
+pluggy==1.3.0
     # via pytest
-pybind11==2.10.3
+pybind11==2.11.1
     # via -r requirements.in
 pybtex==0.24.0
     # via
     #   pybtex-docutils
     #   sphinxcontrib-bibtex
-pybtex-docutils==1.0.2
+pybtex-docutils==1.0.3
     # via sphinxcontrib-bibtex
+pycodestyle==2.11.1
+    # via flake8
 pyflakes==3.1.0
     # via flake8
-pygments==2.15.0
+pygments==2.17.2
     # via sphinx
-pytest==7.2.2
+pytest==7.4.3
     # via
     #   -r requirements.in
     #   pytest-cov
@@ -101,46 +100,45 @@ six==1.16.0
     #   pybtex
 snowballstemmer==2.2.0
     # via sphinx
-soupsieve==2.4.1
+soupsieve==2.5
     # via beautifulsoup4
-sphinx==6.1.3
+sphinx==7.2.6
     # via
     #   myst-parser
     #   sphinx-rtd-theme
     #   sphinx-tippy
+    #   sphinxcontrib-applehelp
     #   sphinxcontrib-bibtex
-sphinx-rtd-theme==1.2.0
+    #   sphinxcontrib-devhelp
+    #   sphinxcontrib-htmlhelp
+    #   sphinxcontrib-jquery
+    #   sphinxcontrib-qthelp
[…]
