+++ requirements.txt
@@ -27,7 +27,7 @@ docutils==0.18.1
     #   sphinxcontrib-bibtex
 exceptiongroup==1.2.0
     # via pytest
-flake8==6.1.0
+flake8==2.3.0
     # via -r requirements.in
 idna==3.6
     # via requests
@@ -64,6 +64,8 @@ packaging==23.2
     # via
     #   pytest
     #   sphinx
+pep8==1.7.1
+    # via flake8
 pluggy==1.3.0
     # via pytest
 pybind11==2.11.1
@@ -74,8 +76,6 @@ pybtex==0.24.0
     #   sphinxcontrib-bibtex
 pybtex-docutils==1.0.3
     # via sphinxcontrib-bibtex
-pycodestyle==2.11.1
-    # via flake8
 pyflakes==3.1.0
     # via flake8
 pygments==2.17.2
