### 2023-11-26 dbf7ba7
Restore compatability libraries
