### 2023-11-26 e21e2a2
Add windows version of optional_throw
