+++ requirements.txt
@@ -2,7 +2,7 @@
 # This file is autogenerated by pip-compile with Python 3.9
 # by the following command:
 #
-#    pip-compile --resolver=backtracking
+#    pip-compile
 #
 alabaster==0.7.13
     # via sphinx
@@ -15,9 +15,7 @@ certifi==2023.11.17
 charset-normalizer==3.3.2
     # via requests
 coverage[toml]==7.3.2
-    # via
-    #   coverage
-    #   pytest-cov
+    # via pytest-cov
 docutils==0.18.1
     # via
     #   myst-parser
@@ -33,8 +31,9 @@ idna==3.6
     # via requests
 imagesize==1.4.1
     # via sphinx
-importlib-metadata==6.8.0
+importlib-metadata==5.2.0
     # via
+    #   -r requirements.in
     #   sphinx
     #   sphinxcontrib-bibtex
 iniconfig==2.0.0
@@ -138,7 +137,11 @@ tomli==2.0.1
     # via
     #   coverage
     #   pytest
+typing-extensions==3.7.4.3
+    # via -r requirements.in
 urllib3==2.1.0
     # via requests
-zipp==3.17.0
-    # via importlib-metadata
+zipp==2.2.1
+    # via
+    #   -r requirements.in
+    #   importlib-metadata
