+++ requirements.txt
@@ -2,7 +2,7 @@
 # This file is autogenerated by pip-compile with Python 3.9
 # by the following command:
 #
-#    pip-compile
+#    pip-compile --resolver=backtracking
 #
 alabaster==0.7.13
     # via sphinx
@@ -15,7 +15,9 @@ certifi==2023.11.17
 charset-normalizer==3.3.2
     # via requests
 coverage[toml]==7.3.2
-    # via pytest-cov
+    # via
+    #   coverage
+    #   pytest-cov
 docutils==0.18.1
     # via
     #   myst-parser
@@ -25,7 +27,7 @@ docutils==0.18.1
     #   sphinxcontrib-bibtex
 exceptiongroup==1.2.0
     # via pytest
-flake8==2.3.0
+flake8==4.0.1
     # via -r requirements.in
 idna==3.6
     # via requests
@@ -51,7 +53,7 @@ markdown-it-py==3.0.0
     #   myst-parser
 markupsafe==2.1.3
     # via jinja2
-mccabe==0.7.0
+mccabe==0.6.1
     # via flake8
 mdit-py-plugins==0.4.0
     # via myst-parser
@@ -63,8 +65,6 @@ packaging==23.2
     # via
     #   pytest
     #   sphinx
-pep8==1.7.1
-    # via flake8
 pluggy==1.3.0
     # via pytest
 pybind11==2.11.1
@@ -75,7 +75,9 @@ pybtex==0.24.0
     #   sphinxcontrib-bibtex
 pybtex-docutils==1.0.3
     # via sphinxcontrib-bibtex
-pyflakes==3.1.0
+pycodestyle==2.8.0
+    # via flake8
+pyflakes==2.4.0
     # via flake8
 pygments==2.17.2
     # via sphinx
