+++ requirements.txt
@@ -2,21 +2,23 @@
 # This file is autogenerated by pip-compile with Python 3.9
 # by the following command:
 #
-#    pip-compile --resolver=backtracking
+#    pip-compile --unsafe-package=numpy --unsafe-package=setuptools requirements.in
 #
 alabaster==0.7.13
     # via sphinx
-babel==2.13.1
+attrs==22.2.0
+    # via pytest
+babel==2.12.1
     # via sphinx
 beautifulsoup4==4.12.2
     # via sphinx-tippy
-certifi==2023.11.17
+certifi==2023.7.22
     # via requests
-charset-normalizer==3.3.2
+charset-normalizer==3.1.0
     # via requests
-coverage[toml]==7.3.2
+coverage[toml]==5.5
     # via
-    #   coverage
+    #   -c constraints.in
     #   pytest-cov
 docutils==0.18.1
     # via
@@ -25,19 +27,16 @@ docutils==0.18.1
     #   sphinx
     #   sphinx-rtd-theme
     #   sphinxcontrib-bibtex
-exceptiongroup==1.2.0
-    # via pytest
-flake8==4.0.1
+flake8==2.3.0
     # via -r requirements.in
-idna==3.6
+idna==3.4
     # via requests
 imagesize==1.4.1
     # via sphinx
 importlib-metadata==5.2.0
     # via
+    #   -c constraints.in
     #   -r requirements.in
-    #   sphinx
-    #   sphinxcontrib-bibtex
 iniconfig==2.0.0
     # via pytest
 jinja2==3.1.2
@@ -47,41 +46,41 @@ jinja2==3.1.2
     #   sphinx-tippy
 latexcodec==2.0.1
     # via pybtex
-markdown-it-py==3.0.0
+markdown-it-py==2.2.0
     # via
     #   mdit-py-plugins
     #   myst-parser
-markupsafe==2.1.3
+markupsafe==2.1.2
     # via jinja2
-mccabe==0.6.1
+mccabe==0.7.0
     # via flake8
-mdit-py-plugins==0.4.0
+mdit-py-plugins==0.3.5
     # via myst-parser
 mdurl==0.1.2
     # via markdown-it-py
-myst-parser==2.0.0
+myst-parser==0.19.1
     # via -r requirements.in
-packaging==23.2
+packaging==23.0
     # via
     #   pytest
     #   sphinx
-pluggy==1.3.0
+pep8==1.7.1
+    # via flake8
+pluggy==1.0.0
     # via pytest
-pybind11==2.11.1
+pybind11==2.10.3
     # via -r requirements.in
 pybtex==0.24.0
     # via
     #   pybtex-docutils
     #   sphinxcontrib-bibtex
-pybtex-docutils==1.0.3
+pybtex-docutils==1.0.2
     # via sphinxcontrib-bibtex
-pycodestyle==2.8.0
+pyflakes==3.1.0
     # via flake8
-pyflakes==2.4.0
-    # via flake8
-pygments==2.17.2
+pygments==2.15.0
     # via sphinx
-pytest==7.4.3
+pytest==7.2.2
     # via
     #   -r requirements.in
     #   pytest-cov
@@ -89,6 +88,7 @@ pytest-cov==4.1.0
     # via -r requirements.in
 pyyaml==6.0.1
     # via
+    #   -c constraints.in
     #   myst-parser
     #   pybtex
 requests==2.31.0
@@ -101,49 +101,46 @@ six==1.16.0
     #   pybtex
 snowballstemmer==2.2.0
     # via sphinx
-soupsieve==2.5
+soupsieve==2.4.1
     # via beautifulsoup4
-sphinx==7.2.6
[…]
