### 2023-11-27 emilydolson/python-phylogeny-tracker#33 @CFGrote
thanks for following up. i usually prefer to check for myself before
closing an issue I opened originally, but no problem here.
