### 2023-11-27 090c859
Add non-wheel numpy test
