### 2023-11-27 ee6f8fc
Add non-wheel numpy test
