### 2023-12-01 ff4cc1b
Pull commented-out np test to @mark.nowheel fn
