### 2023-12-02 cf07ec2
Demonstrate string bug
