+++ test/test_systematics.py
@@ -97,6 +97,83 @@ def test_systematics_numpy():
 
 
 @mark.parametrize(
+    "taxa",
+    (
+        ["hello", "hello 2"],
+        [1, 2],
+        [1.0, 2.0],
+        [[1], [1, 2]],
+    ),
+)
+def test_string_systematics(taxa):
+    tax1, tax2 = taxa
+    sys = systematics.Systematics(lambda x: str(x), True, True, False, False)
+    assert not sys.get_store_position()
+
+    org = tax1
+    org2 = tax2
+    org_tax = systematics.Taxon(0, tax1)
+    org2_tax = sys.add_org(org2, org_tax)
+    org3_tax = sys.add_org(org2)
+    org4_tax = sys.add_org(org, org2_tax)
+    org5_tax = sys.add_org(org, org4_tax)
+
+    assert org_tax.get_info() == str(tax1)
+    assert org2_tax.get_info() == str(tax2)
+    assert org3_tax.get_info() == str(tax2)
+    assert org4_tax.get_info() == str(tax1)
+    assert org5_tax.get_info() == str(tax1)
+    assert org2_tax.get_parent() == org_tax
+    assert org3_tax.get_parent() is None
+    assert org4_tax.get_parent() == org2_tax
+    assert org_tax != org2_tax
+    assert org4_tax != org2_tax
+    assert org5_tax == org4_tax
+    assert sys.get_num_active() == 3
+    assert sys.get_num_ancestors() == 0
+
+    assert not sys.remove_org(org2_tax)
+    assert sys.get_num_active() == 2
+    assert sys.get_num_ancestors() == 1
+
+
+@mark.parametrize(
+    "taxa",
+    (
+        ["hello", "hello 2"],
+        [1, 2],
+        [1.0, 2.0],
+        [[1], [1, 2]],
+    ),
+)
+def test_raw_systematics(taxa):
+    tax1, tax2 = taxa
+    sys = systematics.Systematics()
+    assert not sys.get_store_position()
+
+    org = tax1
+    org2 = tax2
+    org_tax = systematics.Taxon(0, tax1)
+    org2_tax = sys.add_org(org2, org_tax)
+    org3_tax = sys.add_org(org2)
+    org4_tax = sys.add_org(org, org2_tax)
+    org5_tax = sys.add_org(org, org4_tax)
+
+    assert org2_tax.get_parent() == org_tax
+    assert org3_tax.get_parent() is None
+    assert org4_tax.get_parent() == org2_tax
+    assert org_tax != org2_tax
+    assert org4_tax != org2_tax
+    assert org5_tax == org4_tax
+    assert sys.get_num_active() == 3
+    assert sys.get_num_ancestors() == 0
+
+    assert not sys.remove_org(org2_tax)
+    assert sys.get_num_active() == 2
+    assert sys.get_num_ancestors() == 1
+
+
+@mark.parametrize(
     "orgs",
     (
         ["hello", "hello 2"],
@@ -131,6 +208,68 @@ def test_taxa_serialization(orgs):
 def test_numpy_serialization():
     import numpy as np
     test_taxa_serialization([np.array([1]), np.array([1, 2])])
+
+
+@mark.parametrize(
+    "orgs",
+    (
+        ["hello", "hello 2"],
+        [1, 2],
+        [1.0, 2.0],
+        [[1], [1, 2]],
+    ),
+)
+def test_string_taxa_serialization(orgs):
+    org_info_1, org_info_2 = orgs
+    sys = systematics.Systematics(lambda x: str(x))
+    org = org_info_1
+    org2 = org_info_2
+    org_tax = sys.add_org(org)
[…]
