### 2023-12-03 1495c4d
Fix citation syntax
