### 2023-12-03 026fedc
Add new equals operator
