### 2023-12-04 cbd81a9
Check equals operator on construction
