+++ systematics_bindings.cpp
@@ -157,7 +157,12 @@ class taxon_info : public py::object {
 
     void SetEqualsOperator() {
         if (py::isinstance<py::array>(*this)) {
-            equals_operator =  py::module_::import("numpy").attr("array_equal");
+            py::object np = py::module_::import("numpy");
+            if (np) {
+                equals_operator = np.attr("array_equal");
+            } else {
+                equals_operator = this->attr("__class__").attr("__eq__");    
+            }
         } else {
             equals_operator =  this->attr("__class__").attr("__eq__");
         }
