### 2023-12-05 b694881
Fix numpy dependency
