+++ docs/index.md
@@ -4,18 +4,25 @@
 ```
 
 ```{toctree}
+:caption: Phylotrackpy Documentation
 :maxdepth: 3
 :hidden:
 
 introduction
-features
 quickstart
 context
+glossary
+misc
 api
+bibliography
+```
+
+```{toctree}
+:caption: Developer Documentation
+:maxdepth: 3
+:hidden:
+
 contributing
 planning
 community
-```
-
-```{bibliography}
 ```
