+++ requirements.txt
@@ -2,7 +2,7 @@
 # This file is autogenerated by pip-compile with Python 3.9
 # by the following command:
 #
-#    pip-compile
+#    pip-compile --resolver=backtracking
 #
 alabaster==0.7.13
     # via sphinx
@@ -15,7 +15,9 @@ certifi==2023.11.17
 charset-normalizer==3.3.2
     # via requests
 coverage[toml]==7.3.2
-    # via pytest-cov
+    # via
+    #   coverage
+    #   pytest-cov
 docutils==0.20.1
     # via
     #   myst-parser
@@ -31,9 +33,8 @@ idna==3.6
     # via requests
 imagesize==1.4.1
     # via sphinx
-importlib-metadata==5.2.0
+importlib-metadata==7.0.0
     # via
-    #   -r requirements.in
     #   sphinx
     #   sphinxcontrib-bibtex
 iniconfig==2.0.0
@@ -137,11 +138,7 @@ tomli==2.0.1
     # via
     #   coverage
     #   pytest
-typing-extensions==3.7.4.3
-    # via -r requirements.in
 urllib3==2.1.0
     # via requests
-zipp==2.2.1
-    # via
-    #   -r requirements.in
-    #   importlib-metadata
+zipp==3.17.0
+    # via importlib-metadata
