+++ requirements.txt
@@ -4,141 +4,45 @@
 #
 #    pip-compile --resolver=backtracking
 #
-alabaster==0.7.13
-    # via sphinx
-babel==2.13.1
-    # via sphinx
-beautifulsoup4==4.12.2
-    # via sphinx-tippy
-certifi==2023.11.17
-    # via requests
-charset-normalizer==3.3.2
-    # via requests
-coverage[toml]==7.3.2
+attrs==22.2.0
+    # via pytest
+coverage[toml]==5.5
     # via
     #   coverage
     #   pytest-cov
-docutils==0.20.1
-    # via
-    #   myst-parser
-    #   pybtex-docutils
-    #   sphinx
-    #   sphinx-rtd-theme
-    #   sphinxcontrib-bibtex
 exceptiongroup==1.2.0
     # via pytest
-flake8==4.0.1
+flake8==2.3.0
     # via -r requirements.in
-idna==3.6
-    # via requests
-imagesize==1.4.1
-    # via sphinx
-importlib-metadata==7.0.0
-    # via
-    #   sphinx
-    #   sphinxcontrib-bibtex
+importlib-metadata==5.2.0
+    # via -r requirements.in
 iniconfig==2.0.0
     # via pytest
-jinja2==3.1.2
-    # via
-    #   myst-parser
-    #   sphinx
-    #   sphinx-tippy
-latexcodec==2.0.1
-    # via pybtex
-markdown-it-py==3.0.0
-    # via
-    #   mdit-py-plugins
-    #   myst-parser
-markupsafe==2.1.3
-    # via jinja2
-mccabe==0.6.1
+mccabe==0.7.0
     # via flake8
-mdit-py-plugins==0.4.0
-    # via myst-parser
-mdurl==0.1.2
-    # via markdown-it-py
-myst-parser==2.0.0
-    # via -r requirements.in
-packaging==23.2
-    # via
-    #   pytest
-    #   sphinx
-pluggy==1.3.0
+packaging==23.0
     # via pytest
-pybind11==2.11.1
+pep8==1.7.1
+    # via flake8
+pluggy==1.0.0
+    # via pytest
+pybind11==2.10.3
     # via -r requirements.in
-pybtex==0.24.0
-    # via
-    #   pybtex-docutils
-    #   sphinxcontrib-bibtex
-pybtex-docutils==1.0.3
-    # via sphinxcontrib-bibtex
-pycodestyle==2.8.0
+pyflakes==3.1.0
     # via flake8
-pyflakes==2.4.0
-    # via flake8
-pygments==2.17.2
-    # via sphinx
-pytest==7.4.3
+pytest==7.2.2
     # via
     #   -r requirements.in
     #   pytest-cov
 pytest-cov==4.1.0
     # via -r requirements.in
-pyyaml==6.0.1
-    # via
-    #   myst-parser
-    #   pybtex
-requests==2.31.0
-    # via
-    #   sphinx
-    #   sphinx-tippy
-six==1.16.0
-    # via
-    #   latexcodec
-    #   pybtex
-snowballstemmer==2.2.0
-    # via sphinx
-soupsieve==2.5
-    # via beautifulsoup4
-sphinx==7.2.6
-    # via
-    #   myst-parser
-    #   sphinx-rtd-theme
-    #   sphinx-tippy
-    #   sphinxcontrib-applehelp
-    #   sphinxcontrib-bibtex
-    #   sphinxcontrib-devhelp
-    #   sphinxcontrib-htmlhelp
-    #   sphinxcontrib-jquery
-    #   sphinxcontrib-qthelp
-    #   sphinxcontrib-serializinghtml
-sphinx-rtd-theme==2.0.0
[…]
