+++ requirements.in
@@ -1,8 +1,8 @@
 pybind11
-sphinx_rtd_theme
 pytest_cov
-myst_parser>=2.0.0
 pytest
-sphinx_tippy
-sphinxcontrib-bibtex
 flake8<5
+
+typing-extensions<3.8  # for py36 compat
+zipp<3  # for py36 compat
+importlib_metadata<6  # for py36 compat
