+++ requirements.txt
@@ -2,47 +2,145 @@
 # This file is autogenerated by pip-compile with Python 3.9
 # by the following command:
 #
-#    pip-compile --resolver=backtracking
+#    pip-compile --unsafe-package=numpy --unsafe-package=setuptools requirements.in
 #
+alabaster==0.7.13
+    # via sphinx
 attrs==22.2.0
     # via pytest
+babel==2.12.1
+    # via sphinx
+beautifulsoup4==4.12.2
+    # via sphinx-tippy
+certifi==2023.7.22
+    # via requests
+charset-normalizer==3.1.0
+    # via requests
 coverage[toml]==5.5
     # via
-    #   coverage
+    #   -c constraints.in
     #   pytest-cov
-exceptiongroup==1.2.0
-    # via pytest
+docutils==0.18.1
+    # via
+    #   myst-parser
+    #   pybtex-docutils
+    #   sphinx
+    #   sphinx-rtd-theme
+    #   sphinxcontrib-bibtex
 flake8==2.3.0
     # via -r requirements.in
+idna==3.4
+    # via requests
+imagesize==1.4.1
+    # via sphinx
 importlib-metadata==5.2.0
-    # via -r requirements.in
+    # via
+    #   -c constraints.in
+    #   -r requirements.in
 iniconfig==2.0.0
     # via pytest
+jinja2==3.1.2
+    # via
+    #   myst-parser
+    #   sphinx
+    #   sphinx-tippy
+latexcodec==2.0.1
+    # via pybtex
+markdown-it-py==2.2.0
+    # via
+    #   mdit-py-plugins
+    #   myst-parser
+markupsafe==2.1.2
+    # via jinja2
 mccabe==0.7.0
     # via flake8
+mdit-py-plugins==0.3.5
+    # via myst-parser
+mdurl==0.1.2
+    # via markdown-it-py
+myst-parser==0.19.1
+    # via -r requirements.in
 packaging==23.0
-    # via pytest
+    # via
+    #   pytest
+    #   sphinx
 pep8==1.7.1
     # via flake8
 pluggy==1.0.0
     # via pytest
 pybind11==2.10.3
     # via -r requirements.in
+pybtex==0.24.0
+    # via
+    #   pybtex-docutils
+    #   sphinxcontrib-bibtex
+pybtex-docutils==1.0.2
+    # via sphinxcontrib-bibtex
 pyflakes==3.1.0
     # via flake8
+pygments==2.15.0
+    # via sphinx
 pytest==7.2.2
     # via
     #   -r requirements.in
     #   pytest-cov
 pytest-cov==4.1.0
     # via -r requirements.in
+pyyaml==6.0.1
+    # via
+    #   -c constraints.in
+    #   myst-parser
+    #   pybtex
+requests==2.31.0
+    # via
+    #   sphinx
+    #   sphinx-tippy
+six==1.16.0
+    # via
+    #   latexcodec
+    #   pybtex
+snowballstemmer==2.2.0
+    # via sphinx
+soupsieve==2.4.1
+    # via beautifulsoup4
+sphinx==6.1.3
+    # via
+    #   myst-parser
+    #   sphinx-rtd-theme
+    #   sphinx-tippy
+    #   sphinxcontrib-bibtex
+sphinx-rtd-theme==1.2.0
+    # via -r requirements.in
+sphinx-tippy==0.4.1
+    # via -r requirements.in
+sphinxcontrib-applehelp==1.0.4
+    # via sphinx
+sphinxcontrib-bibtex==2.5.0
+    # via -r requirements.in
+sphinxcontrib-devhelp==1.0.2
+    # via sphinx
+sphinxcontrib-htmlhelp==2.0.1
+    # via sphinx
[…]
