+++ .readthedocs.yaml
@@ -28,6 +28,7 @@ sphinx:
 python:
   install:
     - requirements: requirements.txt
+    - requirements: docs/requirements.txt    
     - method: pip
       path: .
 
