+++ requirements.txt
@@ -2,7 +2,7 @@
 # This file is autogenerated by pip-compile with Python 3.9
 # by the following command:
 #
-#    pip-compile --resolver=backtracking
+#    pip-compile --resolver=backtracking --unsafe-package=setuptools requirements.in
 #
 attrs==22.2.0
     # via pytest
