+++ .github/workflows/ci.yml
@@ -27,11 +27,12 @@ jobs:
       uses: actions/setup-python@v4
       with:
         python-version: ${{ matrix.python-version }}
+    - name: Setup LCOV
+      uses: hrishikesh-kadam/setup-lcov@v1        
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
         pip install oldest-supported-numpy==2023.10.25
-        sudo apt install -yy lcov
         # --ignore-requires-python ensures docs deps don't break install on early python versions
         if [ -f requirements.txt ]; then pip install --ignore-requires-python -r requirements.txt; fi
         pip install -e . -Csetup-args=-Db_coverage=true -Cbuilddir=build
