+++ requirements.txt
@@ -4,8 +4,6 @@
 #
 #    pip-compile --resolver=backtracking --unsafe-package=setuptools requirements.in
 #
-attrs==22.2.0
-    # via pytest
 coverage[toml]==7.3.2
     # via
     #   -r requirements.in
@@ -30,7 +28,7 @@ pybind11==2.10.3
     # via -r requirements.in
 pyflakes==3.1.0
     # via flake8
-pytest==7.2.2
+pytest==7.4.3
     # via
     #   -r requirements.in
     #   pytest-cov
