### 2023-12-05 9b47fed
Try without install lcov
