+++ .github/workflows/ci.yml
@@ -27,8 +27,8 @@ jobs:
       uses: actions/setup-python@v4
       with:
         python-version: ${{ matrix.python-version }}
-    - name: Setup LCOV
-      uses: hrishikesh-kadam/setup-lcov@v1        
+    # - name: Setup LCOV
+    #   uses: hrishikesh-kadam/setup-lcov@v1        
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
