### 2023-12-05 cf7a5a4
Try generic lcov install
