+++ .github/workflows/ci.yml
@@ -37,7 +37,10 @@ jobs:
         pip install oldest-supported-numpy==2023.10.25
         # --ignore-requires-python ensures docs deps don't break install on early python versions
         if [ -f requirements.txt ]; then pip install --ignore-requires-python -r requirements.txt; fi
-        pip install -e . -Csetup-args=-Db_coverage=true -Cbuilddir=build
+    - name: Build phylotrackpy 
+      run: |
+        pip install -e . --no-build-isolation -Csetup-args=-Db_coverage=true -Cbuilddir=build
+        ls
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
@@ -47,7 +50,7 @@ jobs:
     - name: Test with pytest
       run: |
         python -m pytest --cov=. --cov-report=lcov
-        lcov --output-file coverage.cpp --capture --directory .
+        lcov --output-file coverage.cpp --capture --directory build
         lcov --output-file coverage.cpp --extract coverage.cpp "*"
     - name: Upload coverage reports to Codecov
       uses: codecov/codecov-action@v3
