+++ .github/workflows/ci.yml
@@ -34,6 +34,7 @@ jobs:
         sudo apt update -yy
         sudo apt install -yy lcov
         python -m pip install --upgrade pip
+        pip install wheel
         pip install oldest-supported-numpy==2023.10.25
         # --ignore-requires-python ensures docs deps don't break install on early python versions
         if [ -f requirements.txt ]; then pip install --ignore-requires-python -r requirements.txt; fi
