+++ docs/conf.py
@@ -71,9 +71,9 @@ author = "Emily Dolson"
 # built documents.
 #
 # The short X.Y version.
-version = "0.1.17"
+version = "0.1.18"
 # The full version, including alpha/beta/rc tags.
-release = "0.1.17"
+release = "0.1.18"
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
