+++ .github/workflows/ci.yml
@@ -27,6 +27,38 @@ jobs:
       uses: actions/setup-python@v4
       with:
         python-version: ${{ matrix.python-version }}
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install oldest-supported-numpy==2023.10.25
+        # --ignore-requires-python ensures docs deps don't break install on early python versions
+        if [ -f requirements.txt ]; then pip install --ignore-requires-python -r requirements.txt; fi
+        pip install -e . 
+    - name: Lint with flake8
+      run: |
+        # stop the build if there are Python syntax errors or undefined names
+        flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
+        # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
+        flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+    - name: Test with pytest
+      run: |
+        python -m pytest
+  
+  coverage:
+    runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        python-version: ["3.11"]
+        os: [ubuntu-latest]
+    steps:
+    - uses: actions/checkout@v4
+      with:
+        submodules: "recursive"
+    - name: Set up Python
+      uses: actions/setup-python@v4
+      with:
+        python-version: ${{ matrix.python-version }}
     # - name: Setup LCOV
     #   uses: hrishikesh-kadam/setup-lcov@v1        
     - name: Install dependencies
@@ -40,14 +72,8 @@ jobs:
         if [ -f requirements.txt ]; then pip install --ignore-requires-python -r requirements.txt; fi
     - name: Build phylotrackpy 
       run: |
-        pip install -e . --no-build-isolation --config-settings setup-args=-Db_coverage=true --config-settings builddir=build
+        make coverage
         ls
-    - name: Lint with flake8
-      run: |
-        # stop the build if there are Python syntax errors or undefined names
-        flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
-        # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-        flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Test with pytest
       run: |
         python -m pytest --cov=. --cov-report=lcov
@@ -57,8 +83,7 @@ jobs:
       uses: codecov/codecov-action@v3
       with:
         files: "coverage.cpp, coverage.lcov"
-        verbose: true
-  
+        verbose: true      
 
   build_wheels:
     name: Build wheels on ${{ matrix.os }}
