### 2024-01-18 05c1d37
Drop python 3.7, fix path
