### 2024-01-18 b8aa8ae
Try to trick codecov into thinking Empirical isnt a submodule
