+++ .github/workflows/ci.yml
@@ -70,6 +70,7 @@ jobs:
         pip install oldest-supported-numpy==2023.10.25
         # --ignore-requires-python ensures docs deps don't break install on early python versions
         if [ -f requirements.txt ]; then pip install --ignore-requires-python -r requirements.txt; fi
+        rm -rf Empirical/.git
     - name: Build phylotrackpy 
       run: |
         make coverage
