+++ requirements.in
@@ -1,8 +1,7 @@
 pybind11
 pytest_cov
-pytest==7.4.3
+pytest
 flake8<5
-coverage>=6.3
 
 typing-extensions<3.8  # for py36 compat
 zipp<3  # for py36 compat
