+++ .github/workflows/ci.yml
@@ -74,7 +74,6 @@ jobs:
     - name: Build phylotrackpy 
       run: |
         make coverage
-        ls
     - name: Test with pytest
       run: |
         python -m pytest --cov=. --cov-report=lcov
