### 2024-01-18 cdebbbb
Try explicitly adding setuptools
