+++ requirements.txt
@@ -2,7 +2,7 @@
 # This file is autogenerated by pip-compile with Python 3.9
 # by the following command:
 #
-#    pip-compile --unsafe-package=numpy --unsafe-package=setuptools requirements.in
+#    pip-compile --unsafe-package=numpy requirements.in
 #
 attrs==22.2.0
     # via pytest
@@ -35,6 +35,8 @@ pytest==7.2.2
     #   -r requirements.in
     #   pytest-cov
 pytest-cov==4.1.0
+    # via -r requirements.in
+setuptools==69.0.3
     # via -r requirements.in
 tomli==2.0.1
     # via pytest
