+++ .github/workflows/ci.yml
@@ -82,7 +82,7 @@ jobs:
     - name: Upload coverage reports to Codecov
       uses: codecov/codecov-action@v3
       with:
-        files: "coverage.cpp, coverage.lcov"
+        files: "coverage.cpp, coverage.lcov, Empirical/include/emp/Evolve/Systematics.hpp"
         verbose: true      
 
   build_wheels:
