+++ codecov.yml
@@ -1,12 +1,3 @@
 codecov: 
  token: f63fa89d-5d9a-4ac8-bfb7-3aac8508ceda
-coverage:
-  status:
-    project:
-      python:
-        paths:
-          - .
-          - ./phylotrackpy*
-      cpp:
-        paths:
-          - Empirical/include/emp/Evolve*
+ disable_default_path_fixes: yes
