### 2024-01-18 fd6bb06
Try rewriting paths
