+++ codecov.yml
@@ -1,3 +1,10 @@
 codecov: 
  token: f63fa89d-5d9a-4ac8-bfb7-3aac8508ceda
  disable_default_path_fixes: yes
+
+ignore:
+  - "usr"
+  - "opt"
+
+fixes:
+  - "home/runner/work/phylotrackpy/*::phylotrackpy/"
