+++ codecov.yml
@@ -5,23 +5,23 @@ codecov:
 ignore:
   - "usr"
   - "opt"
-  - "**/Empirical/include/emp/b*/**"
-  - "**/Empirical/include/emp/c*/**"
-  - "**/Empirical/include/emp/d*/**"
-  - "**/Empirical/include/emp/f*/**"
-  - "**/Empirical/include/emp/g*/**"
-  - "**/Empirical/include/emp/h*/**"
-  - "**/Empirical/include/emp/i*/**"
-  - "**/Empirical/include/emp/m*/**"
-  - "**/Empirical/include/emp/p*/**"
-  - "**/Empirical/include/emp/s*/**"
-  - "**/Empirical/include/emp/t*/**"
-  - "**/Empirical/include/emp/w*/**"
-  - "**/Empirical/include/emp/Evolve/World*.hpp"
-  - "**/Empirical/include/emp/Evolve/NK*.hpp"
-  - "**/Empirical/include/emp/Evolve/O*.hpp"
-  - "**/Empirical/include/emp/Evolve/R*.hpp"
-  - "**/Empirical/include/emp/Evolve/State*.hpp"
+  - "Empirical/include/emp/b*/**"
+  - "Empirical/include/emp/c*/**"
+  - "Empirical/include/emp/d*/**"
+  - "Empirical/include/emp/f*/**"
+  - "Empirical/include/emp/g*/**"
+  - "Empirical/include/emp/h*/**"
+  - "Empirical/include/emp/i*/**"
+  - "Empirical/include/emp/m*/**"
+  - "Empirical/include/emp/p*/**"
+  - "Empirical/include/emp/s*/**"
+  - "Empirical/include/emp/t*/**"
+  - "Empirical/include/emp/w*/**"
+  - "Empirical/include/emp/Evolve/World*.hpp"
+  - "Empirical/include/emp/Evolve/NK*.hpp"
+  - "Empirical/include/emp/Evolve/O*.hpp"
+  - "Empirical/include/emp/Evolve/R*.hpp"
+  - "Empirical/include/emp/Evolve/State*.hpp"
 
 fixes:
   - "home/runner/work/phylotrackpy/*::"
