+++ .github/workflows/ci.yml
@@ -100,7 +100,7 @@ jobs:
         with:
           submodules: "recursive"
       - name: Build wheels
-        uses: pypa/cibuildwheel@v2.16.2
+        uses: pypa/cibuildwheel@v2.16.5
       - uses: actions/upload-artifact@v3
         with:
           path: ./wheelhouse/*.whl
