### 2024-02-29 5a2974c
Skip nonexistent architectures
