+++ joss/paper.bib
@@ -1,4 +1,3 @@
-
 @inproceedings{bohm2017mabe,
   title = {{MABE} (Modular Agent Based Evolver): A framework for digital evolution research},
   author = {Bohm, Clifford and G., Nitash C. and Hintze, Arend},
@@ -26,7 +25,55 @@
   url = {https://doi.org/10.1145/3319619.3326825},
   numpages = 8,
 }
-
+@article{challa2019phylogenetic,
+  title = {Phylogenetic trees: applications, construction, and assessment},
+  author = {Challa, Surekha and Neelapu, Nageswara Rao Reddy},
+  year = 2019,
+  journal = {Essentials of Bioinformatics, Volume III: In Silico Life Sciences: Agriculture},
+  publisher = {Springer},
+  pages = {167--192},
+  doi = {10.1007/978-3-030-19318-8_10},
+}
+@article{clarkeFurtherBiodiversityIndex2001,
+  title = {A further biodiversity index applicable to species lists: variation in taxonomic distinctness},
+  shorttitle = {A further biodiversity index applicable to species lists},
+  author = {Clarke, K. R. and Warwick, R. M.},
+  volume = 216,
+  pages = {265--278},
+  doi = {10.3354/meps216265},
+  issn = {0171-8630, 1616-1599},
+  url = {https://www.int-res.com/abstracts/meps/v216/p265-278/},
+  urldate = {2023-06-10},
+  journaltitle = {Marine Ecology Progress Series},
+  date = {2001-07-06},
+  langid = {english},
+}
+@article{clarkeQuantifyingStructuralRedundancy1998,
+  title = {Quantifying structural redundancy in ecological communities},
+  author = {Clarke, K. R. and Warwick, R. M.},
+  volume = 113,
+  number = 2,
+  pages = {278--289},
+  doi = {10.1007/s004420050379},
+  issn = {1432-1939},
+  url = {https://doi.org/10.1007/s004420050379},
+  urldate = {2023-06-10},
+  journaltitle = {Oecologia},
+  shortjournal = {Oecologia},
+  date = {1998-01-01},
+  langid = {english},
+}
+@article{cock2009biopython,
+  title = {Biopython: freely available Python tools for computational molecular biology and bioinformatics},
+  author = {Cock, Peter JA and Antao, Tiago and Chang, Jeffrey T and Chapman, Brad A and Cox, Cymon J and Dalke, Andrew and Friedberg, Iddo and Hamelryck, Thomas and Kauff, Frank and Wilczynski, Bartek and others},
+  year = 2009,
+  journal = {Bioinformatics},
+  publisher = {Oxford University Press},
+  volume = 25,
+  number = 11,
+  pages = {1422--1423},
+  doi = {10.1093/bioinformatics/btp163},
+}
 @inproceedings{de2012deap,
   title = {{DEAP}: A {Python} Framework for Evolutionary Algorithms},
   author = {De Rainville, Fran\c{c}ois-Michel and Fortin, F\'{e}lix-Antoine and Gardner, Marc-Andr\'{e} and Parizeau, Marc and Gagn\'{e}, Christian},
@@ -57,6 +104,93 @@
   url = {https://doi.org/10.1145/3205651.3205780},
   numpages = 2,
 }
+@article{dolsonDigitalEvolutionEcology2021,
+  title = {Digital Evolution for Ecology Research: A Review},
+  shorttitle = {Digital Evolution for Ecology Research},
+  author = {Dolson, Emily and Ofria, Charles},
+  volume = 9,
+  pages = 852,
+  doi = {10.3389/fevo.2021.750779},
+  issn = {2296-701x},
[…]
