+++ joss/paper.bib
@@ -309,16 +309,22 @@
   doi = {10.21105/joss.04866},
   url = {https://doi.org/10.21105/joss.04866},
 }
-@article{moreno2023lineage,
-  title = {Lineage Tracking Algorithms and the Streaming Curation Problem},
-  author = {Moreno, Matthew Andres and Rodriguez-Papa, Santiago and Dolson, Emily},
-  year = {in review},
-}
 @article{moreno2023toward,
   title = {Toward Phylogenetic Inference of Evolutionary Dynamics at Scale},
   author = {Moreno, Matthew Andres and Dolson, Emily and Rodriguez-Papa, Santiago},
   year = {in press},
   journal = {ALIFE 2023: The 2023 Conference on Artificial Life},
+}
+@misc{moreno2024analysis,
+  title = {Analysis of Phylogeny Tracking Algorithms for Serial and Multiprocess Applications},
+  author = {Matthew Andres Moreno and Santiago Rodriguez Papa and Emily Dolson},
+  year = 2024,
+  publisher = {arXiv},
+  url = {https://arxiv.org/abs/2403.00246},
+  copyright = {arXiv.org perpetual, non-exclusive license},
+  eprint = {2403.00246},
+  archiveprefix = {arXiv},
+  primaryclass = {cs.DS},
 }
 @article{ofria2004avida,
   title = {Avida: A Software Platform for Research in Computational Evolutionary Biology},
