### 2024-03-07 emilydolson/python-phylogeny-tracker#35 @emilydolson
Thanks so much for these suggestions! Profiling info has been added to the paper, we have preprinted the supporting papers, and cited them (thanks @mmore500!). I believe this covers everything.
