+++ joss/paper.bib
@@ -104,11 +104,17 @@
   year={2022}
 }
 
-@article{moreno2023toward,
-  title={Toward Phylogenetic Inference of Evolutionary Dynamics at Scale},
-  author={Moreno, Matthew Andres and Dolson, Emily and Rodriguez-Papa, Santiago},
-  journal={ALIFE 2023: The 2023 Conference on Artificial Life},
-  year={in press}
+@inproceedings{morenoPhylogeneticInferenceEvolutionary2023,
+	title = {Toward Phylogenetic Inference of Evolutionary Dynamics at Scale},
+	url = {https://dx.doi.org/10.1162/isal_a_00694},
+	doi = {10.1162/isal_a_00694},
+	abstract = {Abstract. As digital evolution systems grow in scale and complexity, observing and interpreting their evolutionary dynamics will become increasingly challenging. Distributed and parallel computing, in particular, introduce obstacles to maintaining the high level of observability that makes digital evolution a powerful experimental tool. Phylogenetic analyses represent a promising tool for drawing inferences from digital evolution experiments at scale. Recent work has introduced promising techniques for decentralized phylogenetic inference in parallel and distributed digital evolution systems. However, foundational phylogenetic theory necessary to apply these techniques to characterize evolutionary dynamics is lacking. Here, we lay the groundwork for practical applications of distributed phylogenetic tracking in three ways: 1) we present an improved technique for reconstructing phylogenies from tunably-precise genome annotations, 2) we begin the process of identifying how the signatures of various evolutionary dynamics manifest in phylogenetic metrics, and 3) we quantify the impact of reconstruction-induced imprecision on phylogenetic metrics. We find that selection pressure, spatial structure, and ecology have distinct effects on phylogenetic metrics, although these effects are complex and not always intuitive. We also find that, while low-resolution phylogenetic reconstructions can bias some phylogenetic metrics, high-resolution reconstructions recapitulate them faithfully.},
+	eventtitle = {{ALIFE} 2023: Ghost in the Machine: Proceedings of the 2023 Artificial Life Conference},
+	publisher = {{MIT} Press},
+	author = {Moreno, Matthew Andres and Dolson, Emily and Rodriguez-Papa, Santiago},
+	urldate = {2023-09-27},
+	date = {2023-07-24},
+	langid = {english}
 }
 
 @article{ofria2004avida,
