+++ .github/workflows/ci.yml
@@ -80,7 +80,7 @@ jobs:
         lcov --output-file coverage.cpp --capture --directory .
         lcov --output-file coverage.cpp --extract coverage.cpp "*"
     - name: Upload coverage reports to Codecov
-      uses: codecov/codecov-action@v3
+      uses: codecov/codecov-action@v4
       with:
         files: "coverage.cpp, coverage.lcov, Empirical/include/emp/Evolve/Systematics.hpp"
         verbose: true      
