### 2024-03-07 dccea99
Remove taxon constructor
